update cli for mlops-build and login (#209)(#206)

+++ python/fedml/cli/cli.py
@@ -1,4 +1,5 @@
 import os
+from os.path import expanduser
 
 import click
 import shutil
@@ -80,29 +81,56 @@ def mlops_build(type, source_folder, entry_point, config_folder, dest_folder):
         click.echo("You should specify the type argument value as client or server.")
         exit(-1)
 
+    home_dir = expanduser("~")
+    mlops_build_path = os.path.join(home_dir, "fedml-mlops-build")
+    try:
+        shutil.rmtree(mlops_build_path, ignore_errors=True)
+    except Exception as e:
+        pass
+
+    pip_source_dir = os.path.dirname(__file__)
+    pip_build_path = os.path.join(pip_source_dir, "build-package")
+    shutil.copytree(pip_build_path, mlops_build_path)
+
     if type == "client":
-        build_mlops_package(source_folder, entry_point, config_folder, dest_folder,
-                            "fedml-client", "client-package", "${FEDSYS.CLIENT_INDEX}")
+        result = build_mlops_package(source_folder, entry_point, config_folder, dest_folder,
+                                     mlops_build_path, "fedml-client", "client-package", "${FEDSYS.CLIENT_INDEX}")
+        if result != 0:
+            exit(result)
         click.echo("You have finished all building process. ")
         click.echo("Now you may use " + os.path.join(dest_folder, "client-package.zip") + " to start your federated "
-                   "learning run.")
+                                                                                          "learning run.")
     elif type == "server":
-        build_mlops_package(source_folder, entry_point, config_folder, dest_folder,
-                            "fedml-server", "server-package", "0")
+        result = build_mlops_package(source_folder, entry_point, config_folder, dest_folder,
+                                     mlops_build_path, "fedml-server", "server-package", "0")
+        if result != 0:
+            exit(result)
 
         click.echo("You have finished all building process. ")
         click.echo("Now you may use " + os.path.join(dest_folder, "server-package.zip") + " to start your federated "
-                   "learning run.")
+                                                                                          "learning run.")
 
 
 def build_mlops_package(source_folder, entry_point, config_folder, dest_folder,
-                        mlops_root_dir_name, mlops_package_name, rank):
+                        mlops_build_path, mlops_package_parent_dir, mlops_package_name, rank):
+    if not os.path.exists(source_folder):
+        click.echo("source folder is not exist: " + source_folder)
+        return -1
+
+    if not os.path.exists(os.path.join(source_folder, entry_point)):
+        click.echo("entry file: " + entry_point + " is not exist in the source folder: " + source_folder)
+        return -1
+
+    if not os.path.exists(config_folder):
+        click.echo("config folder is not exist: " + source_folder)
+        return -1
+
     mlops_src = source_folder
     mlops_src_entry = entry_point
     mlops_conf = config_folder
-    cur_dir = os.path.dirname(__file__)
-    mlops_root_dir = os.path.join(cur_dir, "build-package", "mlops-core", mlops_root_dir_name)
-    package_dir = os.path.join(mlops_root_dir, mlops_package_name)
+    cur_dir = mlops_build_path
+    mlops_package_base_dir = os.path.join(cur_dir, "mlops-core", mlops_package_parent_dir)
+    package_dir = os.path.join(mlops_package_base_dir, mlops_package_name)
     fedml_dir = os.path.join(package_dir, "fedml")
     mlops_dest = fedml_dir
     mlops_dest_conf = os.path.join(fedml_dir, "config")
@@ -112,20 +140,27 @@ def build_mlops_package(source_folder, entry_point, config_folder, dest_folder,
     dist_package_dir = os.path.join(dest_folder, "dist-packages")
     dist_package_file = os.path.join(dist_package_dir, mlops_package_file_name)
 
+    shutil.rmtree(mlops_dest_conf, ignore_errors=True)
     shutil.rmtree(mlops_dest, ignore_errors=True)
     shutil.copytree(mlops_src, mlops_dest, copy_function=shutil.copy)
     shutil.copytree(mlops_conf, mlops_dest_conf, copy_function=shutil.copy)
-    os.remove(os.path.join(mlops_dest_conf, "mqtt_config.yaml"))
-    os.remove(os.path.join(mlops_dest_conf, "s3_config.yaml"))
+    try:
+        os.remove(os.path.join(mlops_dest_conf, "mqtt_config.yaml"))
+        os.remove(os.path.join(mlops_dest_conf, "s3_config.yaml"))
+    except Exception as e:
+        pass
 
-    local_mlops_package = os.path.join(mlops_root_dir, mlops_package_file_name)
+    local_mlops_package = os.path.join(mlops_package_base_dir, mlops_package_file_name)
     if os.path.exists(local_mlops_package):
-        os.remove(os.path.join(mlops_root_dir, mlops_package_file_name))
-    mlops_archive_name = os.path.join(mlops_root_dir, mlops_package_name)
-    shutil.make_archive(mlops_archive_name, 'zip', root_dir=mlops_root_dir, base_dir=mlops_package_name)
+        os.remove(os.path.join(mlops_package_base_dir, mlops_package_file_name))
+    mlops_archive_name = os.path.join(mlops_package_base_dir, mlops_package_name)
+    shutil.make_archive(mlops_archive_name, 'zip', root_dir=mlops_package_base_dir, base_dir=mlops_package_name)
     if not os.path.exists(dist_package_dir):
-        os.makedirs(dist_package_dir)
-    if os.path.exists(dist_package_file):
+        try:
+            os.makedirs(dist_package_dir)
+        except Exception as e:
+            pass
+    if os.path.exists(dist_package_file) and not os.path.isdir(dist_package_file):
         os.remove(dist_package_file)
     mlops_archive_zip_file = mlops_archive_name + ".zip"
     if os.path.exists(mlops_archive_zip_file):
@@ -134,7 +169,7 @@ def build_mlops_package(source_folder, entry_point, config_folder, dest_folder,
     mlops_pkg_conf_file = open(mlops_pkg_conf, mode="w")
     mlops_pkg_conf_file.writelines(["entry_config: \n",
                                     "  entry_file: " + mlops_dest_entry + "\n",
-                                    "  conf_file: " + os.path.join("config", "fedml_config.yaml") + "\n"
+                                    "  conf_file: " + os.path.join("config", "fedml_config.yaml") + "\n",
                                     "dynamic_args:\n",
                                     "  rank: " + rank + "\n",
                                     "  run_id: ${FEDSYS.RUN_ID}\n",
@@ -151,6 +186,10 @@ def build_mlops_package(source_folder, entry_point, config_folder, dest_folder,
                                     "  client_num_in_total: ${FEDSYS.CLIENT_NUM}\n"])
     mlops_pkg_conf_file.flush()
     mlops_pkg_conf_file.close()
+
+    shutil.rmtree(mlops_build_path, ignore_errors=True)
+
+    return 0
 
 
 # TODO:
